+++ tests/test_task.py
@@ -9,6 +9,7 @@ import pickle
 import logging
 
 import taskgraph
+import mock
 
 logging.basicConfig(level=logging.DEBUG)
 
@@ -37,7 +38,6 @@ def _div_by_zero():
     """Divide by zero to raise an exception."""
     return 1/0
 
-
 class TaskGraphTests(unittest.TestCase):
     """Tests for the taskgraph."""
 
@@ -51,6 +51,27 @@ class TaskGraphTests(unittest.TestCase):
         """Overriding tearDown function to remove temporary directory."""
         shutil.rmtree(self.workspace_dir)
 
+    def test_version_loaded(self):
+        """TaskGraph: verify we can load the version."""
+        try:
+            import taskgraph
+            # Verifies that there's a version attribute and it has a value.
+            self.assertTrue(len(taskgraph.__version__) > 0)
+        except Exception as error:
+            self.fail('Could not load the taskgraph version as expected.')
+
+    def test_version_not_loaded(self):
+        """TaskGraph: verify exception when not installed."""
+        from pkg_resources import DistributionNotFound
+        import taskgraph
+
+        with mock.patch('taskgraph.pkg_resources.get_distribution',
+                        side_effect=DistributionNotFound('taskgraph')):
+            with self.assertRaises(RuntimeError):
+                # RuntimeError is a side effect of `import taskgraph`, so we
+                # reload it to retrigger the metadata load.
+                taskgraph = reload(taskgraph)
+
     def test_single_task(self):
         """TaskGraph: Test a single task."""
         task_graph = taskgraph.TaskGraph(self.workspace_dir, 0)
@@ -61,6 +82,7 @@ class TaskGraphTests(unittest.TestCase):
             func=_create_list_on_disk,
             args=(value, list_len, target_path),
             target_path_list=[target_path])
+        task_graph.close()
         task_graph.join()
         result = pickle.load(open(target_path, 'rb'))
         self.assertEqual(result, [value]*list_len)
@@ -71,6 +93,7 @@ class TaskGraphTests(unittest.TestCase):
         target_path = os.path.join(self.workspace_dir, '1000.dat')
         _ = task_graph.add_task(
             func=_long_running_function,)
+        task_graph.close()
         timedout = not task_graph.join(0.5)
         # this should timeout since function runs for 5 seconds
         self.assertTrue(timedout)
@@ -97,6 +120,7 @@ class TaskGraphTests(unittest.TestCase):
             func=_create_list_on_disk,
             args=(value, list_len, target_path),
             target_path_list=[target_path])
+        task_graph.close()
         task_graph.join()
 
         # taskgraph shouldn't have recomputed the result
@@ -146,16 +170,70 @@ class TaskGraphTests(unittest.TestCase):
             args=(target_a_path, result_path, result_2_path),
             target_path_list=[result_2_path],
             dependent_task_list=[task_a, sum_task])
+        task_graph.close()
         sum_3_task.join()
         result3 = pickle.load(open(result_2_path, 'rb'))
         expected_result = [(value_a*2+value_b)]*list_len
         self.assertEqual(result3, expected_result)
+        task_graph.join()
+
+
+    def test_task_chain_single_thread(self):
+        """TaskGraph: Test a single threaded task chain."""
+        task_graph = taskgraph.TaskGraph(self.workspace_dir, -1)
+        target_a_path = os.path.join(self.workspace_dir, 'a.dat')
+        target_b_path = os.path.join(self.workspace_dir, 'b.dat')
+        result_path = os.path.join(self.workspace_dir, 'result.dat')
+        result_2_path = os.path.join(self.workspace_dir, 'result2.dat')
+        value_a = 5
+        value_b = 10
+        list_len = 10
+        task_a = task_graph.add_task(
+            func=_create_list_on_disk,
+            args=(value_a, list_len, target_a_path),
+            target_path_list=[target_a_path])
+        task_b = task_graph.add_task(
+            func=_create_list_on_disk,
+            args=(value_b, list_len, target_b_path),
+            target_path_list=[target_b_path])
+        sum_task = task_graph.add_task(
+            func=_sum_lists_from_disk,
+            args=(target_a_path, target_b_path, result_path),
+            target_path_list=[result_path],
+            dependent_task_list=[task_a, task_b])
+        sum_task.join()
+
+        result = pickle.load(open(result_path, 'rb'))
+        self.assertEqual(result, [value_a+value_b]*list_len)
+
+        sum_2_task = task_graph.add_task(
+            func=_sum_lists_from_disk,
+            args=(target_a_path, result_path, result_2_path),
+            target_path_list=[result_2_path],
+            dependent_task_list=[task_a, sum_task])
+        sum_2_task.join()
+        result2 = pickle.load(open(result_2_path, 'rb'))
+        expected_result = [(value_a*2+value_b)]*list_len
+        self.assertEqual(result2, expected_result)
+
+        sum_3_task = task_graph.add_task(
+            func=_sum_lists_from_disk,
+            args=(target_a_path, result_path, result_2_path),
+            target_path_list=[result_2_path],
+            dependent_task_list=[task_a, sum_task])
+        task_graph.close()
+        sum_3_task.join()
+        result3 = pickle.load(open(result_2_path, 'rb'))
+        expected_result = [(value_a*2+value_b)]*list_len
+        self.assertEqual(result3, expected_result)
+        task_graph.join()
 
     def test_broken_task(self):
         """TaskGraph: Test that a task with an exception won't hang."""
         task_graph = taskgraph.TaskGraph(self.workspace_dir, 1)
         _ = task_graph.add_task(
             func=_div_by_zero, task_name='test_broken_task')
+        task_graph.close()
         with self.assertRaises(RuntimeError):
             task_graph.join()
         file_results = glob.glob(os.path.join(self.workspace_dir, '*'))
@@ -176,6 +254,7 @@ class TaskGraphTests(unittest.TestCase):
             args=(value, list_len, target_path),
             target_path_list=[target_path],
             dependent_task_list=[base_task])
+        task_graph.close()
         with self.assertRaises(RuntimeError):
             task_graph.join()
         file_results = glob.glob(os.path.join(self.workspace_dir, '*'))
@@ -186,6 +265,7 @@ class TaskGraphTests(unittest.TestCase):
         """TaskGraph: Test an empty task."""
         task_graph = taskgraph.TaskGraph(self.workspace_dir, 0)
         _ = task_graph.add_task()
+        task_graph.close()
         task_graph.join()
         file_results = glob.glob(os.path.join(self.workspace_dir, '*'))
         # we should have a file in there that's the token
@@ -203,6 +283,7 @@ class TaskGraphTests(unittest.TestCase):
                 func=_create_list_on_disk,
                 args=(value, list_len, target_path),
                 target_path_list=[target_path])
+        task_graph.join()
 
     def test_single_task_multiprocessing(self):
         """TaskGraph: Test a single task with multiprocessing."""
@@ -214,6 +295,7 @@ class TaskGraphTests(unittest.TestCase):
             func=_create_list_on_disk,
             args=(value, list_len, target_path),
             target_path_list=[target_path])
+        task_graph.close()
         task_graph.join()
         result = pickle.load(open(target_path, 'rb'))
         self.assertEqual(result, [value]*list_len)
@@ -240,3 +322,53 @@ class TaskGraphTests(unittest.TestCase):
 
         result = list(_get_file_stats(u'foo', [], False))
         self.assertEqual(result, [])
+
+    def test_encapsulatedtaskop(self):
+        """TaskGraph: Test abstract closure task class."""
+        from taskgraph.Task import EncapsulatedTaskOp
+
+        class TestAbstract(EncapsulatedTaskOp):
+            def __init__(self):
+                pass
+
+        # __call__ is abstract so TypeError since it's not implemented
+        with self.assertRaises(TypeError):
+            x = TestAbstract()
+
+        class TestA(EncapsulatedTaskOp):
+            def __call__(self, x):
+                return x
+
+        class TestB(EncapsulatedTaskOp):
+            def __call__(self, x):
+                return x
+
+        # TestA and TestB should be different because of different class names
+        a = TestA()
+        b = TestB()
+        # results of calls should be the same
+        self.assertEqual(a.__call__(5), b.__call__(5))
+        self.assertNotEqual(a.__name__, b.__name__)
+
+        # two instances with same args should be the same
+        self.assertEqual(TestA().__name__, TestA().__name__)
+
+        # redefine TestA so we get a different hashed __name__
+        class TestA(EncapsulatedTaskOp):
+            def __call__(self, x):
+                return x*x
+
+        new_a = TestA()
+        self.assertNotEqual(a.__name__, new_a.__name__)
+
+        # change internal class constructor to get different hashes
+        class TestA(EncapsulatedTaskOp):
+            def __init__(self, q):
+                super(TestA, self).__init__(q)
+                self.q = q
+
+            def __call__(self, x):
+                return x*x
+
+        init_new_a = TestA(1)
+        self.assertNotEqual(new_a.__name__, init_new_a.__name__)